Revert "Packaging to enable DC2 production using OM10+LensPop"
