+++ PopulationFunctions.py
@@ -1,38 +1,34 @@
-from __future__ import absolute_import, division, print_function
-
-from scipy import interpolate
-from six.moves import cPickle as pickle
-import numpy,math
-import indexTricks as iT
-
 import distances
+from scipy import interpolate
+import cPickle,numpy,math
+import indexTricks as iT
 
 #====================================================================================
 
 class RedshiftDependentRelation():
-
     def __init__(self,D=None,reset=False,cosmo=[0.3,0.7,0.7]):
         self.beginRedshiftDependentRelation(D,reset=reset,cosmo=cosmo)
 
-    def beginRedshiftDependentRelation(self,D,reset,zmax=10,cosmo=[0.3,0.7,0.7]):
+    def beginRedshiftDependentRelation(self,D,reset,zmax=10,cosmo=[0.3,0.7,0.7]):  
         self.zmax=zmax
         self.zbins,self.dz=numpy.linspace(0,self.zmax,401,retstep=True)
         self.z2bins,self.dz2=numpy.linspace(0,self.zmax,201,retstep=True)
         if D==None:
-            D=distance.Distance(cosmo=cosmo)
+            import distances
+            D=distances.Distance(cosmo=cosmo)
         self.D=D
-
+        
         if reset!=True:
             try:
             #load useful redshift splines
                 splinedump=open("redshiftsplines.pkl","rb")
-                self.Da_spline,self.Dmod_spline,self.volume_spline,self.Da_bispline=pickle.load(splinedump)
-            except IOError or EOFError:
+                self.Da_spline,self.Dmod_spline,self.volume_spline,self.Da_bispline=cPickle.load(splinedump)
+            except IOError or EOFError:   
                 self.redshiftfunctions()
         else:
             self.redshiftfunctions()
 
-    def redshiftfunctions(self):
+    def redshiftfunctions(self):   
         D=self.D
         zbins=self.zbins
         z2bins=self.z2bins
@@ -59,7 +55,7 @@ class RedshiftDependentRelation():
 
         #pickle the splines
         splinedump=open("redshiftsplines.pkl","wb")
-        pickle.dump([self.Da_spline,self.Dmod_spline,self.volume_spline,self.Da_bispline],splinedump,2)
+        cPickle.dump([self.Da_spline,self.Dmod_spline,self.volume_spline,self.Da_bispline],splinedump,2)
 
     def Volume(self,z1,z2=None):
         if z2==None:
@@ -74,8 +70,8 @@ class RedshiftDependentRelation():
         elif units=="Mpc":
             corfrac=1
         else:
-            print("don't know those units yet")
-        if z2 is None:
+            print "don't know those units yet"
+        if z2==None:
             return self.splev(z1,self.Da_spline)*corfrac
         else:
             z1,z2=self.biassert(z1,z2)
@@ -124,24 +120,23 @@ class EinsteinRadiusTools(RedshiftDependentRelation):
 
 
 #====================================================================================
-
 class Population(RedshiftDependentRelation):
     def  __init__(self):
         pass
 
-    def draw_apparent_magnitude(self, M, z, band=None, colours=None):
-        if band is not None:
-            colours = self.colour(z,band)
-        if colours is None:
-            colours = 0
-            print("warning no k-correction")
-        Dmods = self.Dmod(z)
+    def draw_apparent_magnitude(self,M,z,band=None,colours=None):
+        if band!=None:
+            colours=self.colour(z,band)
+        if colours==None:
+            colours=0
+            print "warning no k-correction"
+        Dmods=self.Dmod(z)
         ml = M - colours + Dmods
         return ml
-
-    def draw_apparent_size(self, r_phys, z):
-        rl = r_phys/(self.Da(z, units="kpc"))
-        rl *= 206264
+    
+    def draw_apparent_size(self,r_phys,z): 
+        rl = r_phys/(self.Da(z,units="kpc"))
+        rl *= 206264 
         return rl
 
 #====================================================================================
@@ -158,18 +153,18 @@ class LensPopulation_(Population):
         self.beginLensPopulation(D,reset)
 
 
-    def beginLensPopulation(self,D,reset):
+    def beginLensPopulation(self,D,reset): 
         reset=True
         if reset!=True:
             try:
             #load Lens-population splines
                 splinedump=open("lenspopsplines.pkl","rb")
-                self.cdfdNdzasspline,self.cdfdsigdzasspline,self.dNdzspline,self.zlbins,zlmax,sigfloor,self.colourspline,bands=pickle.load(splinedump)
-            except IOError or EOFError or ValueError:
+                self.cdfdNdzasspline,self.cdfdsigdzasspline,self.dNdzspline,self.zlbins,zlmax,sigfloor,self.colourspline,bands=cPickle.load(splinedump)
+            except IOError or EOFError or ValueError:   
                 self.lenspopfunctions()
             #check sigfloor and zlmax are same as requested
             if zlmax!=self.zlmax or self.sigfloor!=sigfloor:
-                self.lenspopfunctions()
+                self.lenspopfunctions() 
             #check all the necessary colours are included
             redocolours=False
             for band in self.bands:
@@ -230,7 +225,7 @@ class LensPopulation_(Population):
 
 
     def Colourspline(self):
-        from stellarpop import tools
+        from stellarpop import tools 
         sed = tools.getSED('BC_Z=1.0_age=10.00gyr')
         #different SEDs don't change things much
 
@@ -238,7 +233,7 @@ class LensPopulation_(Population):
         z=self.zlbins
         self.colourspline={}
         for band in self.bands:
-          if band!="VIS":
+          if band!="VIS":  
             c=z*0
             Cband=tools.filterfromfile(band)
             for i in range(len(z)):
@@ -248,7 +243,7 @@ class LensPopulation_(Population):
 
     def lensPopSplineDump(self):
         splinedump=open("lenspopsplines.pkl","wb")
-        pickle.dump([self.cdfdNdzasspline,self.cdfdNdsigz0asspline,self.cdfdsigdzasspline,self.dNdzspline,self.zlbins,self.zlmax,self.sigfloor,self.colourspline,self.bands],splinedump,2)
+        cPickle.dump([self.cdfdNdzasspline,self.cdfdNdsigz0asspline,self.cdfdsigdzasspline,self.dNdzspline,self.zlbins,self.zlmax,self.sigfloor,self.colourspline,self.bands],splinedump,2)
 
     def draw_z(self,N):
         return interpolate.splev(numpy.random.random(N),self.cdfdNdzasspline)
@@ -260,7 +255,7 @@ class LensPopulation_(Population):
             sigs =interpolate.splev(numpy.random.random(len(z)),self.cdfdNdsigz0asspline)
             return sigs
         else:
-            print("Warning: drawing from 2dpdf is low accuracy")
+            print "Warning: drawing from 2dpdf is low accuracy"
             return self.cdfdsigdzasspline.ev(numpy.random.random(len(z)),z)
 
     def draw_zsig(self,N):
@@ -293,8 +288,8 @@ class LensPopulation_(Population):
             z,zmin=zmin,z
         N=interpolate.splint(zmin,z,self.dNdzspline)
         N*=fsky
-        return N
-
+        return N 
+    
     def phi(self,sigma,z):
         sigma[sigma==0]+=1e-6
         phi_star=(8*10**-3)*self.D.h**3
@@ -333,7 +328,7 @@ class LensPopulation_(Population):
             if band !="VIS":
                 self.ml[band]=self.draw_apparent_magnitude(self.Mr,self.zl,band)
             else: pass
-            self.rl[band]=self.draw_apparent_size(self.r_phys[band],self.zl)
+            self.rl[band]=self.draw_apparent_size(self.r_phys[band],self.zl) 
         return self.zl,self.sigl,self.ml,self.rl,self.ql
 
 #====================================================================================
@@ -343,9 +338,9 @@ class SourcePopulation_(Population):
                   bands=['F814W_ACS','g_SDSS','r_SDSS','i_SDSS','z_SDSS','Y_UKIRT','VIS'],cosmo=[0.3,0.7,0.7],population="cosmos"
                   ):
         self.bands=bands
-
+        
         self.beginRedshiftDependentRelation(D,reset)
-
+        
         if population=="cosmos":
             self.loadcosmos()
         elif population=="lsst":
@@ -357,8 +352,8 @@ class SourcePopulation_(Population):
         try:
             #load pickledcosmos
             cosmosdump=open("cosmosdata.pkl","rb")
-            cosmosphotozs=pickle.load(cosmosdump)
-        except IOError or EOFError:
+            cosmosphotozs=cPickle.load(cosmosdump)
+        except IOError or EOFError:   
             import re
             photozs=open('../Forecaster/cosmos_zphot_mag25.tbl','r').readlines()[10:]
             splinedump=open("cosmosdata.pkl","wb")
@@ -375,7 +370,7 @@ class SourcePopulation_(Population):
             decz=cosmosphotozs[3,:]
             zc=cosmosphotozs[6,:]
             cosmosphotozs=cosmosphotozs[:,((zc<10)&(zc>0))]
-            pickle.dump(cosmosphotozs,splinedump,2)
+            cPickle.dump(cosmosphotozs,splinedump,2)
 
         self.zc=cosmosphotozs[6,:]
 
@@ -400,17 +395,17 @@ class SourcePopulation_(Population):
 
     def loadlsst(self):
         self.population="lsst"
-        import pickle
+        import cPickle
 
         f=open('lsst.1sqdegree_catalog2.pkl','rb')
-        print("new lsst catalogue")
-        data=pickle.load(f)
+        print "new lsst catalogue"
+        data=cPickle.load(f)
         f.close()
-
+        
         self.zc=data[:,2]
         self.m={}
-        #print(data[:,0].max()-data[:,0].min())
-        #print(data[:,1].max()-data[:,1].min())
+        #print data[:,0].max()-data[:,0].min()
+        #print data[:,1].max()-data[:,1].min()
 
         self.m["g_SDSS"]=data[:,3]
         self.m["r_SDSS"]=data[:,4]
@@ -446,7 +441,7 @@ class SourcePopulation_(Population):
         q=q[:N]
 
         return q
-
+     
     def drawSourcePopulation(self,number,sourceplaneoverdensity=10,returnmasses=False):
         source_index=numpy.random.randint(0,len(self.zc),number*3)
         #source_index=source_index[((self.zc[source_index]<10) & (self.zc[source_index]>0.05))]
@@ -463,9 +458,9 @@ class SourcePopulation_(Population):
         self.r_phys=self.RofMz(self.Mvs,self.zs,scatter=True)
         self.rs=self.draw_apparent_size(self.r_phys,self.zs)
         self.qs=self.draw_flattening(number)
-
+        
         self.ps=numpy.random.random_sample(number )*180
-
+        
         #cosmos has a source density of ~0.015 per square arcsecond
         if self.population=="cosmos":
             fac=(0.015)**-0.5
@@ -485,7 +480,7 @@ class SourcePopulation_(Population):
             self.mstar_src=self.mstar[source_index]
             self.mhalo_src=self.mhalo[source_index]
             return self.zs,self.ms,self.xs,self.ys,self.qs,self.ps,self.rs,self.mstar_src,self.mhalo_src
-
+        
         return self.zs,self.ms,self.xs,self.ys,self.qs,self.ps,self.rs
 
 
@@ -495,7 +490,7 @@ class AnalyticSourcePopulation_(SourcePopulation_):
                   ):
         self.bands=bands
         self.beginRedshiftDependentRelation(D,reset)
-        print("not written!")
+        print "not written!"
 
 
 
@@ -503,15 +498,57 @@ if __name__=="__main__":
     #RedshiftDependentRelation(reset=True)
 
     #L=LensPopulation_(reset=True,sigfloor=100)
-
+    
     S=SourcePopulation_(reset=False,population="cosmos")
     S2=SourcePopulation_(reset=False,population="lsst")
 
 
-    print(numpy.median(S.Mv[S.m["i_SDSS"]<25])-numpy.median(S2.Mv[S2.m["i_SDSS"]<25]))
-    print(len(S.Mv[S.m["i_SDSS"]<25])*1./(len(S2.Mv[S2.m["i_SDSS"]<25])*100))
-    print(len(S.Mv)/(60.**2)/2.)
-    print(len(S2.Mv[S2.m["i_SDSS"]<25])/(0.2**2)/(60.**2))
-    print(len(S2.Mv)/(0.2**2)/(60.**2))
+    print numpy.median(S.Mv[S.m["i_SDSS"]<25])-numpy.median(S2.Mv[S2.m["i_SDSS"]<25])
+    print len(S.Mv[S.m["i_SDSS"]<25])*1./(len(S2.Mv[S2.m["i_SDSS"]<25])*100)
+    print len(S.Mv)/(60.**2)/2.
+    print len(S2.Mv[S2.m["i_SDSS"]<25])/(0.2**2)/(60.**2)
+    print len(S2.Mv)/(0.2**2)/(60.**2)
 
-    #print(EarlyTypeRelations(self,100,z=None,scatter=True,band=None))
+    #print EarlyTypeRelations(self,100,z=None,scatter=True,band=None)
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
