+++ imageSim/__init__.py
@@ -1,4 +1,2 @@
-from .convolve import *
-from .pointSource import *
-from .profiles import *
-from .SBmodels import *
+
+
