+++ imageSim/profiles.py
@@ -159,3 +159,4 @@ class _Gauss:
             return s
         except:
             return x*0.
+
