--- lenspop/__init__.py
@@ -1,4 +1,0 @@
-from .population_functions import *
-from .make_lens_pop import *
-from .surveys import *
-from .fast_lens_sim import *
