+++ pylens/MassModels.py
@@ -1,11 +1,4 @@
-from __future__ import absolute_import, division, print_function
-
-from .MassProfiles import PowerLaw as MP_PowerLaw, \
-                          ExtShear as MP_ExtShear, \
-                          PointSource as MP_PointSource, \
-                          Sersic as MP_Sersic, \
-                          SersicG as MP_SersicG, \
-                          sNFW as MP_sNFW
+import MassProfiles
 from math import pi
 
 class MassModel:
@@ -14,7 +7,7 @@ class MassModel:
         self.keys.sort()
         if self.keys not in self._MMkeys:
             import sys
-            print('Not all (or too many) parameters were defined!')
+            print 'Not all (or too many) parameters were defined!'
             sys.exit()
         self._baseProfile.__init__(self)
         self.vmap = {}
@@ -45,8 +38,8 @@ class MassModel:
             self.__setattr__(key,self.vmap[key].value)
 
         
-class PowerLaw(MassModel,MP_PowerLaw):
-    _baseProfile = MP_PowerLaw
+class PowerLaw(MassModel,MassProfiles.PowerLaw):
+    _baseProfile = MassProfiles.PowerLaw
     _MMkeys = [['b','eta','pa','q','x','y'],['b','eta','q','theta','x','y']]
 
     def __init__(self,name,pars):
@@ -59,24 +52,24 @@ class SIE(PowerLaw):
         PowerLaw.__init__(self,name,pars)
 
 
-class ExtShear(MassModel,MP_ExtShear):
-    _baseProfile = MP_ExtShear
+class ExtShear(MassModel,MassProfiles.ExtShear):
+    _baseProfile = MassProfiles.ExtShear
     _MMkeys = [['b','pa','x','y'],['b','theta','x','y']]
 
     def __init__(self,name,pars):
         MassModel.__init__(self,name,pars)
 
 
-class PointSource(MassModel,MP_PointSource):
-    _baseProfile = MP_PointSource
+class PointSource(MassModel,MassProfiles.PointSource):
+    _baseProfile = MassProfiles.PointSource
     _MMkeys = [['b','x','y'],['b','x','y']]
 
     def __init__(self,name,pars):
         MassModel.__init__(self,name,pars)
 
 
-class Sersic(MassModel,MP_Sersic):
-    _baseProfile = MP_Sersic
+class Sersic(MassModel,MassProfiles.Sersic):
+    _baseProfile = MassProfiles.Sersic
     _MMkeys = [['b','n','pa','q','reff','x','y'],
                 ['b','n','pa','q','re','x','y'],
                 ['b','n','q','reff','theta','x','y'],
@@ -101,8 +94,8 @@ class Sersic(MassModel,MP_Sersic):
             self.__dict__[key] = value
 
 
-class SersicG(MassModel,MP_SersicG):
-    _baseProfile = MP_SersicG
+class SersicG(MassModel,MassProfiles.SersicG):
+    _baseProfile = MassProfiles.SersicG
     _MMkeys = [['b','n','pa','q','reff','x','y'],
                 ['b','n','pa','q','re','x','y'],
                 ['b','n','q','reff','theta','x','y'],
@@ -127,8 +120,8 @@ class SersicG(MassModel,MP_SersicG):
             self.__dict__[key] = value
 
 
-class sNFW(MassModel,MP_sNFW):
-    _baseProfile = MP_sNFW
+class sNFW(MassModel,MassProfiles.sNFW):
+    _baseProfile = MassProfiles.sNFW
     _MMkeys = [['b','rs','x','y']]
 
     def __init__(self,name,pars):
