+++ pylens/MassProfiles.py
@@ -1,4 +1,3 @@
-from __future__ import absolute_import, division, print_function
 from ndinterp import ndInterp
 from math import pi
