+++ pylens/__init__.py
@@ -1,4 +1,1 @@
-from .massmodel import *
-from .pylens import *
-from .MassModels import *
-from .models import *
+import massmodel,pylens,MassModels
