+++ pylens/massmodel.py
@@ -1,5 +1,4 @@
-from __future__ import absolute_import, division, print_function
-from .models import _PowerLaw, _ExtShear, _MassSheet, _NFW
+import models
 from math import pi
 
 _PowerLawPars = [['b','eta','pa','q','x','y'],['b','eta','q','theta','x','y']]
@@ -7,7 +6,7 @@ _ExtShearPars = [['b','pa','x','y'],['b','theta','x','y']]
 _MassSheetPars = [['b','x','y'],['b','x','y']]
 _NFWPars = [['b','x','y'],['b','x','y']]
 
-class PowerLaw(_PowerLaw):
+class PowerLaw(models._PowerLaw):
     """
     A subclass for power-law mass models. The `power-law' aspect doesn't
         currently work, but it does work for SIE models.
@@ -22,7 +21,7 @@ class PowerLaw(_PowerLaw):
         keys.sort()
         if keys not in _PowerLawPars:
             import sys
-            print("Not all parameters defined!")
+            print "Not all parameters defined!"
             sys.exit()
         models._PowerLaw.__init__(self)
         self.keys = keys
@@ -64,7 +63,7 @@ class SIE(PowerLaw):
         PowerLaw.__init__(self,name,var,c)
 
 
-class ExtShear(_ExtShear):
+class ExtShear(models._ExtShear):
     def __init__(self,name,var=None,const=None):
         if const is None:
             const = {}
@@ -75,7 +74,7 @@ class ExtShear(_ExtShear):
         keys.sort()
         if keys not in _ExtShearPars:
             import sys
-            print("Not all parameters defined!",keys)
+            print "Not all parameters defined!",keys
             sys.exit()
         models._ExtShear.__init__(self)
         self.keys = keys
@@ -105,7 +104,7 @@ class ExtShear(_ExtShear):
         for key in self.vmap:
             self.__setattr__(key,self.vmap[key].value)
 
-class MassSheet(_MassSheet):
+class MassSheet(models._MassSheet):
     def __init__(self,name,var=None,const=None):
         if const is None:
             const = {}
@@ -116,7 +115,7 @@ class MassSheet(_MassSheet):
         keys.sort()
         if keys not in _MassSheetPars:
             import sys
-            print("Not all parameters defined!",keys)
+            print "Not all parameters defined!",keys
             sys.exit()
         models._MassSheet.__init__(self)
         self.keys = keys
@@ -140,7 +139,7 @@ class MassSheet(_MassSheet):
 
 
 
-class NFW(_NFW):
+class NFW(models._NFW):
     """
     A subclass for power-law mass models. The `power-law' aspect doesn't
         currently work, but it does work for SIE models.
@@ -155,7 +154,7 @@ class NFW(_NFW):
         keys.sort()
         if keys not in _PowerLawPars:
             import sys
-            print("Not all parameters defined!")
+            print "Not all parameters defined!"
             sys.exit()
         models._PowerLaw.__init__(self)
         self.keys = keys
